20.01 Update + 21.01 Hotfix
Pretty much finalised everything I wanted from this framework. A lot of stuff has been completely re-done, it's probably correct to call it Orbinaut Framework 3 at this point. The next update will come with the release of LTS to restore everything that is missing in this version of GameMaker

- Reworked the culling system from the ground up
- Reworked the animator system from the ground up
- Reworked the object collision system from the ground up, now uses built-in collision mask functionalities
- Tweaked tile collision stuff to resolve some issues and improve the workflow
- Tweaked the palette system to simplify it a bit
- Reworked some functions to make them easier to use
- A lot of accuracy tweaks and improvement to player-to-solid collisions
- A lot of player object tweaks to further improve it's accuracy
- A lot of stage object tweaks to further improve their accuracy
- Added native stage transition support via obj_transition_save and obj_transition_load objects
- Improved the way saving system works, allowing to easier implement post-game logic
- Cleaned up the project files from leftover assets
- Resolved a bunch of possible bugs, oversights, inconsistencies etc
- Added Sunset Shore as a playable act*
- Some other changes I simply can't remember because there are A LOT

* Will probably update Sunset Shore to use an actual tileset rather than plain 2048x2048 level sprites to get rid of the texture_prefetch stutter

21.01 hotfix:
- Resolved a crashed caused by incorrect function call (;-;)
- Tweaked SIGNPOST_STATE.CHECK_RESULTS logic
- Fixed a few oversights regarding save file data management in the dev menu

+++ scripts/game_delete_data/game_delete_data.gml
@@ -1,7 +1,10 @@
 /// @self
-/// @description Deletes the saved data for the specified slot.
-/// @param {Real} _slot The slot index of the saved data to delete.
+/// @description		Deletes the saved data for the specified slot.
+/// @param {Real} _slot	The slot index of the saved data to delete.
 function game_delete_data(_slot)
 {
-	file_delete("save" + string(_slot) + ".bin");
+	if _slot != undefined
+	{
+		file_delete("save" + string(_slot) + ".bin");
+	}
 }
